pyproject.toml
[project]
name = "coding-agent"
version = "0.1.0"
description = "An AI coding assistant with explicit Chain-of-Thought reasoning"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "openai>=2.14.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
